requirements.txt
torch~=1.11.0
matplotlib==3.5.2
Pillow~=9.1.1
imageio==2.19.2
numpy==1.22.4
torchvision~=0.12.0
scikit-image==0.19.2